refactor: multi-powered refactor. Added packages structure.
