+++ app/src/main/java/com/ivanmorgillo/corsoandroid/teama/MyApplication.kt
@@ -2,6 +2,8 @@ package com.ivanmorgillo.corsoandroid.teama
 
 import android.app.Application
 import android.os.StrictMode
+import com.ivanmorgillo.corsoandroid.teama.crashlytics.CrashReportingTree
+import com.ivanmorgillo.corsoandroid.teama.crashlytics.LineNumberDebugTree
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
