+++ app/src/main/java/com/ivanmorgillo/corsoandroid/teama/category/Category.kt
@@ -1,0 +1,3 @@
+package com.ivanmorgillo.corsoandroid.teama.category
+
+data class Category(val name: String, val image: String, val id: String)
