+++ app/src/main/java/com/ivanmorgillo/corsoandroid/teama/category/CategoryRepository.kt
@@ -1,16 +1,14 @@
 package com.ivanmorgillo.corsoandroid.teama.category
 
 import com.ivanmorgillo.corsoandroid.teama.network.LoadCategoryResult
-import com.ivanmorgillo.corsoandroid.teama.network.RecipeAPI
+import com.ivanmorgillo.corsoandroid.teama.network.NetworkAPI
 
 interface CategoryRepository {
     suspend fun loadCategories(): LoadCategoryResult
 }
 
-class CategoryRepositoryImpl(private val api: RecipeAPI) : CategoryRepository {
+class CategoryRepositoryImpl(private val api: NetworkAPI) : CategoryRepository {
     override suspend fun loadCategories(): LoadCategoryResult {
         return api.loadCategories()
     }
 }
-
-data class Category(val name: String, val image: String, val id: String)
