+++ app/src/main/java/com/ivanmorgillo/corsoandroid/teama/crashlytics/LineNumberDebugTree.kt
@@ -1,7 +1,8 @@
-package com.ivanmorgillo.corsoandroid.teama
+package com.ivanmorgillo.corsoandroid.teama.crashlytics
 
 import timber.log.Timber
 
+@Suppress("RedundantNullableReturnType")
 class LineNumberDebugTree : Timber.DebugTree() {
     override fun createStackElementTag(element: StackTraceElement): String? {
         return "(${element.fileName}:${element.lineNumber})#${element.methodName}"
