+++ app/src/main/java/com/ivanmorgillo/corsoandroid/teama/detail/DetailViewModel.kt
@@ -3,7 +3,7 @@ package com.ivanmorgillo.corsoandroid.teama.detail
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.ivanmorgillo.corsoandroid.teama.SingleLiveEvent
+import com.ivanmorgillo.corsoandroid.teama.crashlytics.SingleLiveEvent
 import com.ivanmorgillo.corsoandroid.teama.detail.DetailScreenAction.ShowIngredients
 import com.ivanmorgillo.corsoandroid.teama.detail.DetailScreenAction.ShowInstructions
 import com.ivanmorgillo.corsoandroid.teama.detail.DetailScreenAction.ShowInterruptedRequestMessage
@@ -26,7 +26,9 @@ import kotlinx.coroutines.launch
 class DetailViewModel(private val repository: RecipeDetailsRepository) : ViewModel() {
 
     val states = MutableLiveData<DetailScreenStates>()
-    val actions = SingleLiveEvent<DetailScreenAction>()
+    val actions =
+        SingleLiveEvent<DetailScreenAction>()
+
     fun send(event: DetailScreenEvent) {
         when (event) {
             is DetailScreenEvent.OnReady -> {
