+++ app/src/main/java/com/ivanmorgillo/corsoandroid/teama/recipe/RecipeRepository.kt
@@ -1,13 +1,13 @@
 package com.ivanmorgillo.corsoandroid.teama
 
 import com.ivanmorgillo.corsoandroid.teama.network.LoadRecipeResult
-import com.ivanmorgillo.corsoandroid.teama.network.RecipeAPI
+import com.ivanmorgillo.corsoandroid.teama.network.NetworkAPI
 
 interface RecipesRepository {
     suspend fun loadRecipes(categoryName: String): LoadRecipeResult
 }
 
-class RecipeRepositoryImpl(private val api: RecipeAPI) : RecipesRepository {
+class RecipeRepositoryImpl(private val api: NetworkAPI) : RecipesRepository {
     override suspend fun loadRecipes(categoryName: String): LoadRecipeResult {
         return api.loadRecipes(categoryName)
     }
